+++ scripts/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+set -e
+
+npm run prepublish
+./bin/chimp.js "$@"