+++ src/lib/babel-register.js
@@ -1,0 +1,4 @@
+require('babel-register')({
+  'plugins': ['transform-runtime'],
+  'presets': ['es2015', 'stage-2']
+});